feat: Fixed error on wallet lock (#346)
Had to create a starknet.js patch to make it work (as otherwise it errors on the first handler,
and stops sending events for the next ones). The current version disconnects the user cleanly, so
that they can login again easily, but it can be improved with progressive re-login in the future.

Fixes #338