feat: Use polling instead of subscriptions (#600)
I would have loved to use subscriptions, but as of the current state, it is impossible to get event
IDs in a subscription. As it is the only stable id for any given event, and used to avoid duplication
we are forced to use a polling system, that gets the new events every few seconds.